chore: move MSW mocks to "well-known" paths (src/__mocks__)
This should reduce special-cases for configuration files (i.e., tsconfig)
