+++ jest.setup.ts
@@ -1,4 +1,4 @@
-import { server } from "./mocks/server";
+import { server } from "./src/__mocks__/server";
 // Establish API mocking before all tests.
 beforeAll(() => server.listen());
