+++ src/__mocks__/server.ts
@@ -1,5 +1,5 @@
-import { setupServer } from "msw/node";
-import { handlers } from "./handlers";
+import {setupServer} from 'msw/node';
+import {handlers} from './handlers';
 
 // This configures a request mocking server with the given request handlers.
 export const server = setupServer(...handlers);
