+++ src/lib/services/globus-connect-server/__tests__/endpoint.spec.ts
@@ -1,7 +1,7 @@
 import { createStorage } from "../../../core/storage";
 import { endpoint } from "..";
 
-import type { MirroredRequest } from "../../../../../mocks/handlers";
+import type { MirroredRequest } from "../../../../__mocks__/handlers";
 
 const GCS_HOST = "https://fa5e.bd7c.data.globus.org";
