+++ src/lib/services/globus-connect-server/__tests__/user-credentials.spec.ts
@@ -1,7 +1,7 @@
 import { createStorage } from "../../../core/storage";
 import { userCredentials } from "..";
 
-import type { MirroredRequest } from "../../../../../mocks/handlers";
+import type { MirroredRequest } from "../../../../__mocks__/handlers";
 
 const GCS_HOST = "https://fa5e.bd7c.data.globus.org";
