+++ src/lib/services/groups/__tests__/groups.spec.ts
@@ -1,7 +1,7 @@
 import { groups } from "..";
 import { createStorage } from "../../../core/storage";
 
-import type { MirroredRequest } from "../../../../../mocks/handlers";
+import type { MirroredRequest } from "../../../../__mocks__/handlers";
 
 test("get", async () => {
   createStorage("memory");
