+++ src/lib/services/groups/__tests__/policies.spec.ts
@@ -1,7 +1,7 @@
 import { policies } from "..";
 import { createStorage } from "../../../core/storage";
 
-import type { MirroredRequest } from "../../../../../mocks/handlers";
+import type { MirroredRequest } from "../../../../__mocks__/handlers";
 
 test("policies - get", async () => {
   createStorage("memory");
