+++ src/lib/services/timer/__tests__/timer.spec.ts
@@ -1,7 +1,7 @@
 import { create } from "../index";
 import { createStorage } from "../../../core/storage";
 
-import type { MirroredRequest } from "../../../../../mocks/handlers";
+import type { MirroredRequest } from "../../../../__mocks__/handlers";
 
 test("create", async () => {
   createStorage("memory");
